upgrade horizonclient to patched version to fix load offers issues (closes #369) (#373)
* Revert "Upgrade horizonclient to v2 to fix horizon v1 API changes (closes #360) (#362)"

This reverts commit ba198426b99e7919a16ec998503ec5d0143d38bf.

* 2 - update glide.yaml and glide.lock file to use v2.0 of stellar/go repo for the horizonclient dependency

* 3 - use forked version of stellar/go at nikhilsaraf/go branch "horizonclient-v2.0.0/1_export_offerBase"

* 4 - upgrade to second patched version of horizonclient (fee stats fix)

+++ plugins/sdexExtensions.go
@@ -85,7 +85,7 @@ func getFeeFromStats(horizonClient horizonclient.ClientInterface, capacityTrigge
 	return maxOpFeeStroops, nil
 }
 
-func getMaxFee(fs *hProtocol.FeeStats, percentile uint8) (int64, error) {
+func getMaxFee(fs *hProtocol.FeeStats, percentile uint8) (int, error) {
 	switch percentile {
 	case 10:
 		return fs.MaxFee.P10, nil
